commit changes for form and help functions
