+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "jest"
+    "test": "jest  --coverage"
   },
   "dependencies": {
     "@jest/globals": "^29.5.0",
