+++ src/pages/ChargerMap/ChargerMap.js
@@ -1,7 +1,9 @@
 import Script from 'next/script';
 import {Helmet} from 'react-helmet';
 // import Filters from './filters.js';
-import Filters2 from './filters2.js';
+import FiltersForm from '../../../components/FiltersForm.js';
+import getConnectionsFilters from '../../../helper_functions/getConnectionsFilters.js';
+import getOperatorsFilters from '../../../helper_functions/getOperatorsFilters.js';
 // import activity list
 import ActivityList from 'src/pages/ActivityList/ActivityList.js';
 
@@ -27,30 +29,12 @@ export default function ChargerMap(props) {
     connections:[]
   });
 
- function getOperatorsFilters (filterArray) {
-   let ofilters = [];
-   if(filterArray.operators.length!==0) {
-    ofilters=filters.operators.map((operator)=>{
-      return ['in', operator, ['string', ['get', 'poi']]];
-     });
-   }
-   return ofilters;
- }
 
- function getConnectionsFilters (filterArray) {
-  let cfilters = [];
-  if(filterArray.connections.length!==0) {
-    cfilters = filters.connections.map((connection)=>{
-      return ['in', connection, ['string', ['get', 'connectionType']]];
-  });
 
-  };
 
-  return cfilters;
-}
 
 function handleClick() {
-  console.log("in handle click");
+  // console.log("in handle click");
   let combinedFilters = getOperatorsFilters(filters).concat(getConnectionsFilters(filters));
   if (combinedFilters.length!== 0) {
     let filter = ['any',].concat(combinedFilters);
@@ -192,7 +176,7 @@ function jsonEscape(str)  {
  Longitude: {lng} | Latitude: {lat} | Zoom: {zoom}
  </div> */}
   {/* <div> */}
-     <Filters2 filters={filters} setFilters={setFilters} onCloseClick={handleClick} />
+     <FiltersForm filters={filters} setFilters={setFilters} onCloseClick={handleClick} />
      <div ref={mapContainer} className="map-container" />
   </span>
   <pre id="quake-info">
