+++ src/styles/globals.css
@@ -202,7 +202,7 @@ mapboxgl-popup {
   padding-bottom: 6px;
   text-align: center;
   font-size: 16px;
-  margin: 4px 2px;
+  margin: 4px 4px;
   border-radius: 30%;
 
 }
